Add endpoints for adding and deleting chats
Added a POST and DELETE endpoint.

The POST endpoint uses service discovery to connect to the
uniborrow-users-service and checks, if the user ids are valid. If they
are, it will save the chat. The string message is sanitizied.

The delete endpoint takes a chat id as path parameter to delete the
chat.